Changed migration to make it more reliable

+++ lib/active_record/migration/app_generator.rb
@@ -920,6 +920,9 @@ module ActiveRecord
         to_model_class_name = ref_config[:to_model_class_name]
         to_no_master = ref_config[:no_master_association]
         from_what = ref_config[:from]
+
+        return if Admin::MigrationGenerator.view_exists?("#{schema}.#{ref_view_name(to_table_name)}")
+
         <<~DO_TEXT
           DROP VIEW if exists #{schema}.#{ref_view_name(to_table_name)};
           CREATE VIEW #{schema}.#{ref_view_name(to_table_name)} AS
